Move Outermost target selector feature to inspector_main
We realized we were getting console errors when running DevTools in node
debugging mode (node_app). The errors where about SDK models used by the
outermost target selector feature not being available, which is expected
because the ispected target is node.

To resolve this, we moved the feature registration from main to
inspector_main, which isn't part of the node_app.

Fixed: 1457203
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4637321
Commit-Queue: Nancy Li <nancyly@chromium.org>
(cherry picked from commit 9d5feea9350e67271be028acc3fd0f7d734e5d5c)
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4640222
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Nancy Li <nancyly@chromium.org>
