+++ config/gni/devtools_grd_files.gni
@@ -811,15 +811,15 @@ grd_files_debug_sources = [
   "front_end/entrypoints/heap_snapshot_worker/HeapSnapshotLoader.js",
   "front_end/entrypoints/heap_snapshot_worker/HeapSnapshotWorkerDispatcher.js",
   "front_end/entrypoints/inspector_main/InspectorMain.js",
+  "front_end/entrypoints/inspector_main/OutermostTargetSelector.js",
   "front_end/entrypoints/inspector_main/RenderingOptions.js",
   "front_end/entrypoints/inspector_main/nodeIcon.css.js",
+  "front_end/entrypoints/inspector_main/outermostTargetSelector.css.js",
   "front_end/entrypoints/inspector_main/renderingOptions.css.js",
   "front_end/entrypoints/lighthouse_worker/LighthouseWorkerService.js",
   "front_end/entrypoints/main/ExecutionContextSelector.js",
   "front_end/entrypoints/main/MainImpl.js",
-  "front_end/entrypoints/main/OutermostTargetSelector.js",
   "front_end/entrypoints/main/SimpleApp.js",
-  "front_end/entrypoints/main/outermostTargetSelector.css.js",
   "front_end/entrypoints/node_app/NodeConnectionsPanel.js",
   "front_end/entrypoints/node_app/NodeMain.js",
   "front_end/entrypoints/node_app/nodeConnectionsPanel.css.js",
