+++ front_end/entrypoints/inspector_main/inspector_main.ts
@@ -6,9 +6,11 @@ import './RenderingOptions.js';
 import './InspectorMain.js';
 
 import * as InspectorMain from './InspectorMain.js';
+import * as OutermostTargetSelector from './OutermostTargetSelector.js';
 import * as RenderingOptions from './RenderingOptions.js';
 
 export {
   InspectorMain,
+  OutermostTargetSelector,
   RenderingOptions,
 };
