+++ front_end/entrypoints/inspector_main/outermostTargetSelector.css
