+++ front_end/entrypoints/main/BUILD.gn
@@ -4,18 +4,12 @@
 
 import("../../../scripts/build/ninja/devtools_entrypoint.gni")
 import("../../../scripts/build/ninja/devtools_module.gni")
-import("../../../scripts/build/ninja/generate_css.gni")
 import("../visibility.gni")
-
-generate_css("css_files") {
-  sources = [ "outermostTargetSelector.css" ]
-}
 
 devtools_module("main") {
   sources = [
     "ExecutionContextSelector.ts",
     "MainImpl.ts",
-    "OutermostTargetSelector.ts",
     "SimpleApp.ts",
   ]
 
@@ -45,10 +39,7 @@ devtools_module("main") {
 devtools_entrypoint("bundle") {
   entrypoint = "main.ts"
 
-  deps = [
-    ":css_files",
-    ":main",
-  ]
+  deps = [ ":main" ]
 
   visibility = [
     ":*",
