+++ front_end/entrypoints/main/main.ts
@@ -4,12 +4,10 @@
 
 import * as ExecutionContextSelector from './ExecutionContextSelector.js';
 import * as MainImpl from './MainImpl.js';
-import * as OutermostTargetSelector from './OutermostTargetSelector.js';
 import * as SimpleApp from './SimpleApp.js';
 
 export {
   ExecutionContextSelector,
   MainImpl,
-  OutermostTargetSelector,
   SimpleApp,
 };
